# agent/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/v1/device/getED713Ota": {
            "post": {
                "description": "get ED713 ota version information",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ED713"
                ],
                "summary": "getED713Ota",
                "parameters": [
                    {
                        "description": "OTA request",
                        "name": "in",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mdb.Ed713OtqReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mysql.Ed713OtaData"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "mdb.Ed713OtqReq": {
            "type": "object",
            "properties": {
                "base_version": {
                    "type": "integer"
                },
                "core_version": {
                    "type": "integer"
                },
                "mac": {
                    "type": "string"
                }
            }
        },
        "mysql.Ed713OtaData": {
            "type": "object",
            "properties": {
                "baseFileSize": {
                    "type": "integer"
                },
                "baseOtaUrl": {
                    "type": "string"
                },
                "coreFileSize": {
                    "type": "integer"
                },
                "coreOtaUrl": {
                    "type": "string"
                },
                "mac": {
                    "type": "string"
                },
                "remoteBaseVersion": {
                    "type": "integer"
                },
                "remoteCoreVersion": {
                    "type": "integer"
                },
                "upgrade": {
                    "type": "integer"
                }
            }
        }
    }
}